Updated practiceNests...
I answer the second question and actually made the first question work, A side note when working with if statements, you have to be very specific with the order they go in. Going in different orders changed the results if statements only pay attention to the if statement in front of them.